+++ src/main/java/insane96mcp/insanelib/base/LoadFeature.java
@@ -1,7 +1,16 @@
 package insane96mcp.insanelib.base;
 
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE})
 public @interface LoadFeature {
     String module();
+    String name() default "";
+    String description() default "";
     boolean enabledByDefault() default true;
     boolean canBeDisabled() default true;
     String[] requiresMods() default "";