+++ src/main/java/insane96mcp/insanelib/base/config/Config.java
@@ -8,6 +8,8 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
 public @interface Config {
+    String name() default "";
+    String description() default "";
     double min() default -Double.MAX_VALUE;
     double max() default Double.MAX_VALUE;
 }