+++ src/main/java/insane96mcp/insanelib/module/base/BaseFeature.java
@@ -1,11 +1,9 @@
 package insane96mcp.insanelib.module.base;
 
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 
-@Label(name = "Base")
 @LoadFeature(module = "insanelib:base", canBeDisabled = false)
 public class BaseFeature extends Feature {